+++ pyodim/__init__.py
@@ -1,6 +1,5 @@
 # primary read routines
 from .pyodim import read_odim
-from .pyodim import read_write_odim
 
 # helper routines
 from .pyodim import copy_h5_data